+++ src/network_ops_dashboard/asavpn/scripts/showvpnconnected.py
@@ -1,7 +1,6 @@
 import requests
 import logging
 from requests.auth import HTTPBasicAuth
-from network_ops_dashboard.models import SiteSecrets, NetworkCredential
 from network_ops_dashboard.inventory.models import Inventory
 from network_ops_dashboard.asavpn.models import AsaVpnConnectedUsers 
 
@@ -10,16 +9,11 @@ logger = logging.getLogger('network_ops_dashboard.asavpn')
 def showVPNconnected():
     #urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
     headers = {'User-Agent': 'ASDM'}
-    try:
-        creds_name = SiteSecrets.objects.filter(varname='asavpn_primary_user')[0].varvalue
-        creds = NetworkCredential.objects.filter(username_lookup=creds_name)
-    except ImportError as e:
-        logger.exception(f"No asavpn_primary_user set in SiteSecrets.objects(): {e}")
-    auth = HTTPBasicAuth(creds[0].username, creds[0].password)
     baseStr = '/admin/exec/show%20vpn-sessiondb'
     asaInv = Inventory.objects.filter(device_tag__name__exact='ASAVPN') # Update tag name if necessary or remove tags to ignore
     for TargetDevice in asaInv:
         try:
+            auth = HTTPBasicAuth(TargetDevice.creds_rest.username, TargetDevice.creds_rest.password)
             r = requests.get('https://' + str(TargetDevice) + baseStr, headers=headers, auth=auth, verify=False)
             if r.text != '':
                     try: