+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 
+## v0.2.2 - 08-15-2025
+- ⚙️ On-call: Added ability to auto-archive incidents that are "Closed" daily or weekly at a provided time.
+- ✅ On-call: Added status "Archived". Incident log will now show incidents that are "Archived".
+- ✅ Scripts: Lookup NetworkCredential from Inventoy object instead of referring to it due to duplicates and ease.
+- 🗃️ Inventory.object (WIP): Add Forgeignkey fields to NetworkCredential for SSH and REST.
+- 🖥️ UX: Added superuser views to edit/add SiteSettings & SiteSettingsWebsites.
+- 🖥️ UX: Added staff views to edit/add SiteSecrets to function with network scripts.
+- 🖥️ UX: Added net-admin views to edit/add NetworkCredential & added new model field "username_search_field"
+
 ## v0.2.1 - 08-15-2025
 - 🖥️ UX: Added styling to all buttons across the project.
 - ⚙️ Filters: Added filtering to Iventory, Sites, Reports/Changes, Reports/Circuits, & Reports/Circuits/Circuit
 - 🆕 On-call: Added email settings, automated cronjob enablement, & module configuration modal.
-
 
 ## v0.2.0 - 08-14-2025
 - 🆕 Dashboard: Added customize button to arrange & toggle display of cards. Staff users can also enable/disable scripted cards & change script settings.